// frontend/src/shared/api/generated/model/getCollectionItemsShelfOfShame200DataItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Boardgames tracker API
 * API для проекта Boardgames tracker
 * OpenAPI spec version: 1.0.0
 */
import type { GetCollectionItemsShelfOfShame200DataItemBoardGame } from './getCollectionItemsShelfOfShame200DataItemBoardGame';
import type { GetCollectionItemsShelfOfShame200DataItemStatus } from './getCollectionItemsShelfOfShame200DataItemStatus';

export type GetCollectionItemsShelfOfShame200DataItem = {
  readonly id: number;
  readonly userId: number;
  boardGame: GetCollectionItemsShelfOfShame200DataItemBoardGame;
  /** @nullable */
  datePurchased?: string | null;
  /**
   * @minimum 0
   * @nullable
   */
  sumInRubles?: number | null;
  status: GetCollectionItemsShelfOfShame200DataItemStatus;
  /** @minimum 0 */
  readonly playCount: number;
  /**
   * @maxLength 2000
   * @nullable
   */
  comment?: string | null;
  readonly createdAt: string;
  readonly updatedAt: string;
};